requirements.txt
django==3.2.6
djangorestframework
django-cors-headers
Pillow